Improving the NavBar
