+++ webapp/src/App.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
-import NavBar from "./components/NavBar";
+import NavBar from "./components/navigation/NavBar";
 import DedeApp from "./components/DedeApp";
 import Home from "./components/Home";
 import Shop from "./components/Shop";
